get tags from postgres, #340
