--- server/handlers/orgs/browse-add-tags-view.js
@@ -1,20 +1,0 @@
-'use strict';
-
-var helpers = require('../helpers.js');
-
-module.exports = function (request, reply) {
-  var permissions = helpers.getPermissions(request.auth.credentials);
-  var pageType = 'orgs';
-  var options = {};
-
-  request.pg.tags.getAllActive(function (error, allActive) {
-    options.pageType = pageType;
-    options.allActive = allActive;
-    options.permissions = permissions;
-    if (Object.keys(options.allActive).length === 0) {
-      return reply.view('browse/no-tags', options);
-    }
-
-    return reply.view('browse/categories-tags', options);
-  });
-};
