+++ server/handlers/shared/browse-add-tags-view.js
@@ -1,0 +1,21 @@
+'use strict';
+
+var helpers = require('../helpers.js');
+
+module.exports = function (pageType) {
+  return function (request, reply) {
+    var permissions = helpers.getPermissions(request.auth.credentials);
+    var options = {};
+
+    request.pg.tags.getAllActive(function (error, allActive) {
+      options.pageType = pageType;
+      options.allActive = allActive;
+      options.permissions = permissions;
+      if (Object.keys(options.allActive).length === 0) {
+        return reply.view('browse/no-tags', options);
+      }
+
+      return reply.view('browse/categories-tags', options);
+    });
+  };
+};
