--- templates/views/browse/tags.html
@@ -1,27 +1,0 @@
-<!-- Include Top Navbar -->
-{{>top-nav}}
-
-<!-- Main Content -->
-<div class="container">
-  <div class="card card--header">
-    <h1 class="title">Select Tags</h1>
-  </div>
-  {{#if parent_tags}}
-    <form class="form form--tags" method="post">
-        <div class="form__section">
-          <label class="tag__title">
-            <a class="anchor--danger text-decoration" href="/{{searchAll}}">Search all tags</a>
-          </label>
-
-            {{#each parent_tags}}
-              {{>filter-tag view=../view}}
-            {{/each}}
-        </div>
-    </form>
-  {{else}}
-    <h3>Sorry! No tags were found</h3>
-  {{/if}}
-</div>
-
-<!-- Include Navbar -->
-{{>navbar}}
